refactor(contracts): split domain request models
